backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==4.2.19
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
gitdb==4.0.11
GitPython==3.1.41
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.22.3
setuptools==69.0.3
smmap==5.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1